scsi: ufs: Fix power management OPs handling
Currently, the ufshcd driver does forceful resume and suspend
from system_suspend if device is runtime suspended. Next, when
the system resume is invoked, it finds that the runtime status is
still suspended and bails out. This leads to the runtime PM
framework throwing errors when we try to set the runtime status
of the children as active. This leads to the hardware going into
a bad state.
Fix this by doing a forceful resume as a part of system_resume
and also marking the runtime status of ufshcd dev as ACTIVE.

Change-Id: Id4454a2f32f3980ad7f8df62815fba4e5ce43ebe
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: atndko <z1281552865@gmail.com>